refactor#8 devide components more and value props
