+++ src/components/features/view-page/PostAreaContainer.jsx
@@ -1,46 +1,24 @@
-import React from 'react';
-import Button from '@commons/Button';
-import { MdFavoriteBorder, MdFavorite, MdKeyboard, MdOutlineModeComment, MdModeComment } from 'react-icons/md';
-import { FaTrashAlt, FaEdit } from 'react-icons/fa';
-import { CgProfile } from 'react-icons/cg';
+import { MdFavoriteBorder, MdFavorite, MdOutlineModeComment, MdModeComment } from 'react-icons/md';
 import styled from 'styled-components';
 import IconButton from '@/components/commons/IconButton';
+import PostInfomationContainer from './in-post-area-container/PostInfomationContainer';
+import PostOnAuthButtons from './in-post-area-container/PostOnAuthButtons';
 
-export default function PostAreaContainer() {
+export default function PostAreaContainer({ PostWriter }) {
   return (
     <StPostAreaContainer>
-      <StPostTitleWrapper>
-        <h1>넵</h1>
-      </StPostTitleWrapper>
-      <StLanguageTypeWrapper>
-        <MdKeyboard size={30} />
-        <p>JavaScript</p>
-      </StLanguageTypeWrapper>
-      <StPostWriterWrapper>
-        <CgProfile size={30} />
-        <p>익명의 작성자</p>
-      </StPostWriterWrapper>
-      <StCodeBoxWrapper>
-        <p>코드박스</p>
-      </StCodeBoxWrapper>
+      <PostInfomationContainer PostWriter={PostWriter} />
       <StPostToggleButtonContainer>
         <StPostLikeButtonContainer>
-          <StIconButton activeIcon={MdFavoriteBorder} inActiveIcon={MdFavorite} />
-          <p>999</p>
+          <StLikeButton activeIcon={MdFavoriteBorder} inActiveIcon={MdFavorite} />
+          <p>{PostWriter.Likes}</p>
         </StPostLikeButtonContainer>
         <StCommentToggleButtonContainer>
-          <StIconButton activeIcon={MdOutlineModeComment} inActiveIcon={MdModeComment} />
-          <p>99</p>
+          <StCommentButton activeIcon={MdOutlineModeComment} inActiveIcon={MdModeComment} />
+          <p>{PostWriter.Comments}</p>
         </StCommentToggleButtonContainer>
       </StPostToggleButtonContainer>
-      <StPostEditButtonContainer>
-        <StButton $size="sm" $variant="solid" $color="point">
-          <FaEdit />
-        </StButton>
-        <StButton $size="sm" $variant="solid" $color="point">
-          <FaTrashAlt />
-        </StButton>
-      </StPostEditButtonContainer>
+      {PostWriter.IsAuth ? <PostOnAuthButtons /> : <></>}
     </StPostAreaContainer>
   );
 }
@@ -62,56 +40,7 @@ const StPostToggleButtonContainer = styled.div`
   justify-self: center;
 `;
 
-const StPostEditButtonContainer = styled.div`
-  grid-area: 9 / 1 / 9 / 2;
-  display: grid;
-  grid-auto-flow: column;
-  column-gap: 12px;
-  justify-self: center;
-`;
-
-const StPostTitleWrapper = styled.div`
-  grid-area: 1 / 1 / 2 / 7;
-  text-align: center;
-  font-size: var(--font-size-lg);
-
-  & h1 {
-    background-color: var(--color-point-light);
-    border-radius: var(--round-md) var(--round-md) 0px 0px;
-    padding: 15px 15px 15px 15px;
-  }
-`;
-
-const StLanguageTypeWrapper = styled.div`
-  grid-area: 2 / 6 / 3 / 5;
-  text-align: center;
-`;
-
-const StPostWriterWrapper = styled.div`
-  grid-area: 2 / 6 / 3 / 7;
-  text-align: center;
-`;
-
-const StCodeBoxWrapper = styled.div`
-  margin: 14px 20px 0px 20px;
-  padding: 18px 18px 18px 18px;
-  background-color: var(--color-white);
-  border: 1px solid var(--color-border);
-  border-radius: var(--round-xl);
-  grid-area: 3 / 1 / 9 / 7;
-  text-align: start;
-  font-size: var(--font-size-md);
-`;
-
-const StButton = styled(Button)`
-  width: 38px;
-  height: 38px;
-  color: black;
-  background-color: transparent;
-  border: 1px solid transparent;
-`;
-
-const StIconButton = styled(IconButton)`
+const StLikeButton = styled(IconButton)`
   font-size: 28px;
   background-color: transparent;
   border: none;
@@ -121,10 +50,20 @@ const StIconButton = styled(IconButton)`
   }
 `;
 
+const StCommentButton = styled(IconButton)`
+  font-size: 28px;
+  background-color: transparent;
+  border: none;
+  color: var(--color-black);
+  &:hover {
+    background-color: transparent;
+  }
+`;
+
 const StCommentToggleButtonContainer = styled.div`
   justify-self: center;
   text-align: center;
-  font-size: var(--font-size-sm);
+  font-size: var(--font-size-md);
 `;
 
 const StPostLikeButtonContainer = styled.div`
